JDP200501-2 GroupController wstepna implementacja endpointow v3
