+++ src/main/java/com/kodilla/ecommercee/domain/GroupDto.java
@@ -2,12 +2,16 @@ package com.kodilla.ecommercee.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @AllArgsConstructor
+@NoArgsConstructor
 @Getter
+@Setter
 public class GroupDto {
 
-    Long id;
-    String name;
+    private Long groupId;
+    private String name;
 
 }
